requirements.txt
fastapi
uvicorn
pandas
Pillow
pydantic
python-dotenv
openai
dashscope
thefuzz
pypinyin
requests